Refactored the AltitudeMethod flow into its own package.
