+++ src/main/java/gov/acwi/wqp/etl/altitudeMethod/AltitudeMethodProcessor.java
@@ -1,6 +1,8 @@
-package main.java.gov.acwi.wqp.etl;
+package main.java.gov.acwi.wqp.etl.altitudeMethod;
 
 import org.springframework.batch.item.ItemProcessor;
+
+import main.java.gov.acwi.wqp.etl.GwReflist;
 
 public class AltitudeMethodProcessor implements ItemProcessor<GwReflist, AltitudeMethod> {
