+++ src/main/java/gov/acwi/wqp/etl/altitudeMethod/DeleteAltitudeMethod.java
@@ -1,4 +1,4 @@
-package main.java.gov.acwi.wqp.etl;
+package main.java.gov.acwi.wqp.etl.altitudeMethod;
 
 import org.springframework.batch.core.StepContribution;
 import org.springframework.batch.core.configuration.annotation.StepScope;
